+++ modules/alloy/src/main/ts/ephox/alloy/behaviour/disabling/DisableTypes.ts
@@ -13,7 +13,7 @@ export interface DisableBehaviour extends Behaviour.AlloyBehaviour<DisableConfig
 }
 
 export interface DisableConfig extends Behaviour.BehaviourConfigDetail {
-  disabled: boolean;
+  disabled: () => boolean;
   disableClass: Option<string>;
   useNative: boolean;
   onEnabled: (comp: AlloyComponent) => void;
@@ -21,7 +21,7 @@ export interface DisableConfig extends Behaviour.BehaviourConfigDetail {
 }
 
 export interface DisableConfigSpec extends Behaviour.BehaviourConfigSpec {
-  disabled?: boolean;
+  disabled?: () => boolean;
   disableClass?: string;
   useNative?: boolean;
   onEnabled?: (comp: AlloyComponent) => void;